public/agent_requirements.txt
pyserial>=3.5
pyautogui>=0.9.54
Pillow>=10.0.0
pygetwindow>=0.0.9
colorama>=0.4.6
requests>=2.31.0
